Adding a test runner and sqlite settings. Updating and cleaning tests.
I have redone CSRF tests to be self-contained. I rather not have a test_app
for testing the app, which is not included when installing django-uni-form,
so running `./manage.py test uni_form` fails

I have removed some comments that I find unnecessary, removed some duplicates
and cleaned a little bit the code.

The runner easily allows to test the application, without needing a project
or having to specify an specific settings file.

I have added a new test, that checks CSRF token is not displayed when using
GET on a form.

More tests coming soon :)